Enable package plugins with a 5.6 deployment target without a special environment variable
This removes the environment variable and sets the availability annotations to 5.6.
