+++ Fixtures/Miscellaneous/Plugins/SandboxTesterPlugin/Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version: 5.5
+// swift-tools-version: 5.6
 import PackageDescription
 
 let package = Package(
