+++ Tests/PackageLoadingTests/PD_5_6_LoadingTests.swift
@@ -9,6 +9,7 @@
 */
 
 import PackageModel
+import TSCBasic
 import XCTest
 
 class PackageDescription5_6LoadingTests: PackageDescriptionLoadingTests {
@@ -66,4 +67,51 @@ class PackageDescription5_6LoadingTests: PackageDescriptionLoadingTests {
         }
     }
 
+    func testBuildToolPluginTarget() throws {
+        let stream = BufferedOutputByteStream()
+        stream <<< """
+            import PackageDescription
+            let package = Package(
+               name: "Foo",
+               targets: [
+                   .plugin(
+                       name: "Foo",
+                       capability: .buildTool()
+                    )
+               ]
+            )
+            """
+
+        loadManifest(stream.bytes) { manifest in
+            XCTAssertEqual(manifest.targets[0].type, .plugin)
+            XCTAssertEqual(manifest.targets[0].pluginCapability, .buildTool)
+        }
+    }
+
+    func testPluginTargetCustomization() throws {
+        let stream = BufferedOutputByteStream()
+        stream <<< """
+            import PackageDescription
+            let package = Package(
+               name: "Foo",
+               targets: [
+                   .plugin(
+                       name: "Foo",
+                       capability: .buildTool(),
+                       path: "Sources/Foo",
+                       exclude: ["IAmOut.swift"],
+                       sources: ["CountMeIn.swift"]
+                    )
+               ]
+            )
+            """
+
+        loadManifest(stream.bytes) { manifest in
+            XCTAssertEqual(manifest.targets[0].type, .plugin)
+            XCTAssertEqual(manifest.targets[0].pluginCapability, .buildTool)
+            XCTAssertEqual(manifest.targets[0].path, "Sources/Foo")
+            XCTAssertEqual(manifest.targets[0].exclude, ["IAmOut.swift"])
+            XCTAssertEqual(manifest.targets[0].sources, ["CountMeIn.swift"])
+        }
+    }
 }
